pursuit-domain/requirements-python36.txt
yaaf==1.0.3
gym==0.21.0
pygame
pyyaml
mlflow
